Arreglo en el modulo de creación de organos con cargas masivas. Look and feel.

+++ compras/protected/controllers/UsuariosController.php
@@ -205,17 +205,18 @@ public function actionUpdate($id)
 */
 public function actionDelete($id)
 {
-if(Yii::app()->request->isPostRequest)
-{
-// we only allow deletion via POST request
-$this->loadModel($id)->delete();
+	if(Yii::app()->request->isPostRequest)
+	{
+		// we only allow deletion via POST request
+		if($this->loadModel($id)->esHijo())
+		$this->loadModel($id)->delete();
 
-// if AJAX request (triggered by deletion via admin grid view), we should not redirect the browser
-if(!isset($_GET['ajax']))
-$this->redirect(isset($_POST['returnUrl']) ? $_POST['returnUrl'] : array('admin'));
-}
-else
-throw new CHttpException(400,'Invalid request. Please do not repeat this request again.');
+		// if AJAX request (triggered by deletion via admin grid view), we should not redirect the browser
+		if(!isset($_GET['ajax']))
+			$this->redirect(isset($_POST['returnUrl']) ? $_POST['returnUrl'] : array('admin'));
+	}
+	else
+		throw new CHttpException(400,'Invalid request. Please do not repeat this request again.');
 }
 
 /**
